requirements.txt
torch
torchvision
pillow
numpy
scikit-learn
matplotlib
seaborn
opencv-python
tqdm